5 - SQLI/Ch1 - WhiteBox Injection/requirements.txt
flask==3.0.1
mysql-connector-python==8.0.33
